Forgot to add the updated cpp file with the new changes

+++ src/modules/math_logic.cpp
@@ -1,25 +1,26 @@
 module;
 #include <cmath>
-#include <numbers> // std::numbers::pi
+#include <cstdint>
+#include <numbers>
 #include <vector>
 
 module math_logic;
 
-void NumberGenerator::AddPoint(unsigned long long p) {
+void NumberGenerator::AddPoint(uint64_t p) {
   double r = static_cast<double>(p);
   double theta = std::fmod(static_cast<double>(p), 2.0 * std::numbers::pi);
 
   NumberPoint pt;
-  pt.x = r * std::cos(theta);
-  pt.y = -r * std::sin(theta);
-  pt.p = p;
+  pt.x = static_cast<float>(r * std::cos(theta));
+  pt.y = static_cast<float>(-r * std::sin(theta));
+  pt.p = static_cast<float>(p);
 
-  points.push_back(pt); // adds to end of std::vector
+  points.push_back(pt);
 }
 
 void NumberGenerator::GeneratePrimesInRange(int rangeSize) {
-  unsigned long long low = lastChecked + 1;
-  unsigned long long high = low + rangeSize - 1;
+  uint64_t low = lastChecked + 1;
+  uint64_t high = low + rangeSize - 1;
 
   // Initializes the list
   std::vector<char> isNotPrime(rangeSize, 0);
@@ -27,9 +28,9 @@ void NumberGenerator::GeneratePrimesInRange(int rangeSize) {
   // If on first sieve
   if (low == 1) {
     isNotPrime[0] = 1; // 1 is not prime
-    for (unsigned long long p = 2; p * p <= high; p++) {
+    for (uint64_t p = 2; p * p <= high; p++) {
       if (!isNotPrime[p - low]) {
-        for (unsigned long long j = p * p; j <= high; j += p)
+        for (uint64_t j = p * p; j <= high; j += p)
           isNotPrime[j - low] = 1;
       }
     }
@@ -37,12 +38,12 @@ void NumberGenerator::GeneratePrimesInRange(int rangeSize) {
 
   // Sieves based on already calculated primes
   for (const auto &pt : points) {
-    unsigned long long p = pt.p;
+    uint64_t p = pt.p;
     if (p * p > high)
       break;
 
     // Finds first multiple of p in range (low, high) truncating the division
-    unsigned long long start = (low / p) * p;
+    uint64_t start = (low / p) * p;
     if (start < low)
       start += p;
 
@@ -51,7 +52,7 @@ void NumberGenerator::GeneratePrimesInRange(int rangeSize) {
       start = p * p;
 
     // Marks multiples of p as non-prime
-    for (unsigned long long j = start; j <= high; j += p) {
+    for (uint64_t j = start; j <= high; j += p) {
       isNotPrime[j - low] = 1;
     }
   }
@@ -66,13 +67,13 @@ void NumberGenerator::GeneratePrimesInRange(int rangeSize) {
 }
 
 void NumberGenerator::GenerateMultiplesInRange(unsigned int n, int rangeSize) {
-  unsigned long long low = lastChecked + 1;
-  unsigned long long high = low + rangeSize - 1;
+  uint64_t low = lastChecked + 1;
+  uint64_t high = low + rangeSize - 1;
 
   // Finds the first multiple of n starting from 'low'
-  unsigned long long start = ((low + n - 1) / n) * n;
+  uint64_t start = ((low + n - 1) / n) * n;
 
-  for (unsigned long long i = start; i <= high; i += n) {
+  for (uint64_t i = start; i <= high; i += n) {
     AddPoint(i);
   }
   lastChecked = high;
